refactor: update OwnerInvoiceDTO and OwnerInvoiceMapper to include email field
