--- core/src/main/java/com/dam/commune/invoice/Untitled-1.sql
@@ -1,1 +1,0 @@
-select * from invoice;
