+++ core/src/main/java/com/dam/commune/microservices/pdfservice/PdfService.java
@@ -1,4 +1,4 @@
-package com.dam.commune.invoice;
+package com.dam.commune.microservices.pdfservice;
 
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.MediaType;
